Evaluate parameters

Given documented parameter settings settings, a test matrix test_mat, PCP solution pcp_model, and binary mask test_mask, returns statistics about the pcp_model's performance with settings on the test_mat. This is an internal function needed by grid_search_cv(). It is not expected that users should require access to this function.

Usage

eval_params(settings, test_mat, pcp_model, test_mask)

Arguments

settings

Documented parameter settings associated with the given pcp_model.

test_mat

The given test matrix with test_mask observed (i.e. not missing).

pcp_model

The output of a PCP algorithm, either rrmc() or root_pcp().

test_mask

The binary mask indicating which entries in test_mat comprise the test set.

Value

A data.frame object with PCP's performance statistics.

See also