Estimate rank of a given matrix

matrix_rank() estimates the rank of a given data matrix D by counting the number of "practically nonzero" singular values of D.

A singular value \(s\) is determined to be "practically nonzero" if \(s \geq s_max * thresh\), i.e. if it is greater than or equal to the maximum singular value in D scaled by a given threshold thresh.

Usage

matrix_rank(D, thresh = NULL)

Arguments

D

The input data matrix.

thresh

(Optional) A double eqn{> 0}, specifying the relative threshold by which "practically zero" is determined, used to calculate the rank of D. By default, thresh = NULL, in which case the threshold is set to max(dim(D)) * .Machine$double.eps.

Value

An integer representing the rank of D.

See also

Examples

data <- sim_data()
matrix_rank(data$D)
#> [1] 10
matrix_rank(data$L)
#> [1] 3