Proximal gradient method for the Frobenius norm

prox_frobenius() implements the proximal gradient method for the Frobenius norm. The proximal gradient method is used to solve non-differentiable convex optimization problems. In root_pcp(), this thresholding minimizes the square root of the sum of squared error, where the error is defined as Z = D - L - S. This is an internal function needed by root_pcp(). It is not expected that users should require access to this function.

Usage

prox_frobenius(Z, c)

Arguments

Z

The input error/noise matrix, Z = D - L - S.

c

The amount by which the prox Frobenius method penalizes Z.

Value

The thresholded error/noise matrix.

See also