Compute singular values of given matrix

sing() calculates the singular values of a given data matrix D. This is done with a call to svd(), and is included in pcpr to enable the quick characterization of a data matrix's raw low-rank structure, to help decide whether rrmc() or root_pcp() is the more appropriate PCP algorithm to employ in conjunction with D.

Experimentally, the rrmc() approach to PCP has best been able to handle those datasets that are governed by complex underlying patterns characterized by slowly decaying singular values, such as EH data. For observed data with a well-defined low rank structure (rapidly decaying singular values), root_pcp() may offer a better model estimate.

Usage

sing(D)

Arguments

D

The input data matrix.

Value

A numeric vector containing the singular values of D.

References

"Singular value decomposition" Wikipedia article.

Examples

data <- sim_data()
sing(data$D)
#>  [1] 27.3359578  2.8782309  2.2675947  1.1444730  0.9990220  0.5674336
#>  [7]  0.5121203  0.4584272  0.4516461  0.4085756
# could plot the singular values for visual inspection with e.g.
# plot(sing(data$D), type = 'b')