Estimate sparsity of given matrix

sparsity() estimates the percentage of entries in a given data matrix D whose values are "practically zero". If the absolute value of an entry is below a given threshold parameter thresh, then that value is determined to be "practically zero", increasing the estimated sparsity of D.

Usage

sparsity(D, thresh = 1e-04)

Arguments

D

The input data matrix.

thresh

(Optional) A numeric threshold >= 0 used to determine if an entry in D is "practically zero". If the absolute value of an entry is below thresh, then it is judged to be "practically zero". By default, thresh = 1e-04.

Value

The sparsity of D, measured as the percentage of entries in D that are "practically zero".

See also

Examples

sparsity(matrix(rep(c(1, 0), 8), 4, 4))
#> [1] 50
sparsity(matrix(0:8, 3, 3))
#> [1] 11.11111
sparsity(matrix(0, 3, 3))
#> [1] 100